refactor: remove serial module and update WlTransys struct for serialization
